Create tcs online cards.c++
A group of friends are playing cards. Only numeric cards are in consideration along 
with the joker card (equivalent to 0) 12 the symbols in the cards 
are not taken into account. Each player will receive a set of 
cards. The rule of game is to rearrange the set of cards to the 
possible number sequence. Player with the set of cards forming the 
smallest number wins the game. The sequence of cards should not 
start with a joker card. Task is to write a program for developing
the logic of the game considering the set of cards as a number.

Example 1: 
Input: 
3421 -> Value of Card Set
Output:
1234 -> Prints the lowest number matching the set criteria

Example 2:

Input:

4278-> Value of Card Set Output:

Output:

2478-> Prints the lowest number matching the set